Change NEEDED_REGISTERS to u8 for keccak and bigint

+++ jolt-inlines/keccak256/src/test_utils.rs
@@ -22,7 +22,7 @@ pub type KeccakTestCase = InstructionTestCase<Keccak256State, Keccak256State>;
 /// Wrapper around `CpuTestHarness` that offers convenient Keccak helpers.
 pub struct KeccakCpuHarness {
     pub harness: CpuTestHarness,
-    pub vr: [u8; NEEDED_REGISTERS],
+    pub vr: [u8; NEEDED_REGISTERS as usize],
 }
 
 impl KeccakCpuHarness {
@@ -35,7 +35,7 @@ impl KeccakCpuHarness {
         let guards: Vec<_> = (0..NEEDED_REGISTERS)
             .map(|_| tracer::utils::virtual_registers::allocate_virtual_register_for_inline())
             .collect();
-        let vr: [u8; NEEDED_REGISTERS] = core::array::from_fn(|i| *guards[i]);
+        let vr: [u8; NEEDED_REGISTERS as usize] = core::array::from_fn(|i| *guards[i]);
 
         Self {
             harness: CpuTestHarness::new(),
